pyproject.toml
[project]
name = "dailo-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.11.1",
    "boto3>=1.42.91",
    "celery>=5.6.3",
    "certifi==2026.2.25",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "charset-normalizer==3.4.7",
    "click==8.3.2",
    "cloudinary==1.44.1",
    "daphne>=4.2.1",
    "django==6.0.4",
    "django-cloudinary-storage==0.3.0",
    "django-cors-headers==4.9.0",
    "django-environ==0.13.0",
    "django-filter==25.2",
    "django-storages>=1.14.6",
    "djangorestframework==3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "h11==0.16.0",
    "idna==3.11",
    "pillow==12.2.0",
    "psycopg2-binary==2.9.11",
    "pyjwt==2.12.1",
    "redis==7.4.0",
    "requests==2.33.1",
    "six==1.17.0",
    "sqlparse==0.5.5",
    "stripe==15.0.1",
    "typing-extensions==4.15.0",
    "urllib3==2.6.3",
    "uvicorn[standard]==0.44.0",
    "websockets>=16.0",
    "whitenoise==6.12.0",
]
